bump version of aws sdk from 1.9.25 to 0.10.0
