+++ plugin/project/Build.scala
@@ -4,20 +4,9 @@ import sbtrelease.ReleasePlugin._
 
 object Build extends sbt.Build {
 
-  val playVersion = "2.2.1"
   val ScalaVersion ="2.10.3"
   val libName = "s3-play-plugin"
   val libOrganization = "org.corespring"
-
-  object Dependencies {
-    val play = "com.typesafe.play" %% "play" % playVersion % "provided"
-    val playTest = "com.typesafe.play" %% "play-test" % playVersion % "test"
-    val aws = "com.amazonaws" % "aws-java-sdk" % "1.9.25"
-    val config = "com.typesafe" % "config" % "1.0.0"
-    val akkaTest = "com.typesafe.akka" %% "akka-testkit" % "2.1.0" % "test"
-    val scalaTest = "org.scalatest" % "scalatest_2.10" % "1.9.1" % "test"
-    val all = Seq(scalaTest, play, playTest, aws, config, akkaTest)
-  }
 
   object Resolvers {
     val typesafe = "Typesafe Releases" at "http://repo.typesafe.com/typesafe/releases/"
